requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
elevenlabs==1.2.2
google-cloud-storage==3.4.0
google-auth==2.41.1